% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{evidence.raw}
\alias{evidence.raw}
\title{evidence file from MaxQuant of a TMT experiment of MDA-MB231 cell
protrusions and bodies during a timecourse experiment.
The dataset contains the proteinGroup.txt output file from
\href{http://www.maxquant.org}{MaxQuant}.}
\format{
A dataframe with 5905 rows and 107 variables.
The important variables are as follows:

\itemize{
  \item{Sequence}{The identified AA sequence of the peptide.}
  \item{Length}{The length of the sequence stored in the column 'Sequence'.}
  \item{Modifications}{Post-translational modifications contained within the
  identified peptide sequence.}
  \item{Raw.file}{The name of the RAW-file the mass spectral data was derived from.}
  \item{Score}{Andromeda score for the best associated MS/MS spectrum.}
}
#' @return A data.frame
}
\source{
Dermit et al 2020.
Subcellular mRNA Localization Regulates Ribosome Biogenesis in Migrating Cells
Developmental Cell Volume 55, Issue 3, 9 November 2020, Pages 298-313.e10.
\href{https://www.cell.com/developmental-cell/pdfExtended/S1534-5807(20)30796-6}.
}
\usage{
evidence.raw
}
\description{
evidence file from MaxQuant of a TMT experiment of MDA-MB231 cell
protrusions and bodies during a timecourse experiment.
The dataset contains the proteinGroup.txt output file from
\href{http://www.maxquant.org}{MaxQuant}.
}
\examples{
head(evidence.raw)
}
\keyword{datasets}
