% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_MaxQuant.R
\name{filter_MaxQuant}
\alias{filter_MaxQuant}
\title{Filter proteinGroups table by rows}
\usage{
filter_MaxQuant(proteins, tofilter)
}
\arguments{
\item{proteins}{Data.frame,
Protein table originating from MaxQuant.}

\item{tofilter}{Character,
Name of the column(s) containing features to be filtered on.}
}
\value{
dataframe
}
\description{
\code{filter_MaxQuant} imports a proteinGroups table from MaxQuant
and filter rows with a '+' on columns to filter provided by user.
}
\examples{
if(interactive()){
proteins_filtered <- filter_MaxQuant(prot.raw,
tofilter= c("Reverse" , "Potential.contaminant" ,"Only.identified.by.site"))

}
}
