% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_TMT_se.R
\name{make_TMT_se}
\alias{make_TMT_se}
\title{Data.frame to SummarizedExperiment object
conversion using file information and user input}
\usage{
make_TMT_se(
  proteins_unique,
  columns_positions,
  intensities,
  time_unit = 30,
  time_span = c(1, 2, 4, 8, 16),
  numerator = "prot",
  denominator = "body",
  sep = "_"
)
}
\arguments{
\item{proteins_unique}{Data.frame,
Protein table with unique names annotated in the 'name' column
(output from \code{\link{make_unique}()}).}

\item{columns_positions}{Integer,
position of columns that contain experiment quantitative data}

\item{intensities}{Character,
names of columns that contain experiment quantitative data}

\item{time_unit}{Integer,
unit of time in that defines the experiment}

\item{time_span}{Integer,
number of times that the time_unit is repeated at the experimental timepoint}

\item{numerator}{Character,
condition of interest (prot)}

\item{denominator}{character,
condition to make relative (body)}

\item{sep}{Character,
The separator used to parse the column header}
}
\value{
A SummarizedExperiment object
with log2-transformed values, normalised to control and median subtracted
(by column).
}
\description{
\code{make_TMT_se} creates a SummarizedExperiment object
based on protein table and user's input about experimental design.
}
\examples{
if(interactive()){
# Load example
data <- prot.raw
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+" &
data$Reverse != "+" ,]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

columns_positions<-str_which(colnames(data_unique),
"Reporter.intensity.corrected.(\\\\d)+.(\\\\d)")
intensities <- colnames(data_unique)[str_which(colnames(data_unique),
"Reporter.intensity.corrected.(\\\\d)+.(\\\\d)")]
# Make SummarizedExperiment
se <- make_TMT_se(data_unique,columns_positions,intensities,
time_unit=30,
time_span=c(1,2,4,8,16),
numerator= "prot",
denominator= "body",
sep = "_")
}
}
