% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{plot_scatter}
\usage{
plot_scatter(se, xpos, ypos, category, color_highlight, xlimit, ylimit)
}
\arguments{
\item{se}{SummarizedExperiment,
Data object containing processed proteomics data}

\item{xpos}{Integer,
Timepoint to represent on the x axis on the scatterplot}

\item{ypos}{Integer,
Timepoint to represent on the y axis on the scatterplot}

\item{category}{Character,
Sets the proteins to highlight in the plot}

\item{color_highlight}{Character,
Sets the color of category to highlight in the plot}

\item{xlimit}{Character,
Sets the limit of x axis on the scatterplot}

\item{ylimit}{Character,
Sets the limit of y axis on the scatterplot}
}
\value{
A scatter plot (generated by \code{\link[ggplot2]{ggplot}})
}
\description{
plot_scatter
}
\examples{
if(interactive()){
plot_scatter(se, 1, 2, "HIST", 'orange', 4, 4)
}
}
