% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_1D.R
\name{enrich_1D}
\alias{enrich_1D}
\title{enrich_1D}
\usage{
enrich_1D(
  se,
  timepoint = 5,
  dbs = "GO_Molecular_Function_2018",
  number_dbs = 1
)
}
\arguments{
\item{se}{SummarizedExperiment,
Data object containing processed proteomics data}

\item{timepoint}{integer
Timepoint to do enrichment analysis}

\item{dbs}{Character
Character vector of databases to search}

\item{number_dbs}{integer}
}
\value{
A tibble
with enrichment terms, p-values etc
}
\description{
enrich_1D
}
\examples{
if(interactive()){
enrichment_table <- enrich_1D(se,timepoint= 5, dbs ="GO_Molecular_Function_2018",number_dbs=1)
}
}
