% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peptides.raw}
\alias{peptides.raw}
\title{peptides file from MaxQuant of a TMT experiment of MDA-MB231 cell
protrusions and bodies during a timecourse experiment.}
\format{
A data.frame with 50603 observations and 12 variables:
\describe{
  \item{Sequence}{The identified AA sequence of the peptide}
  \item{Amino.acid.before}{The amino acid in the sequence before the peptide}
  \item{Missed.cleavages}{Number of missed enzymatic cleavages}
  \item{Mass}{Monoisotopic mass of the peptide}
  \item{Proteins}{Identifiers of proteins this peptide is associated with}
}
}
\source{
Dermit et al 2020.
Subcellular mRNA Localization Regulates Ribosome Biogenesis in Migrating Cells.
Developmental Cell Volume 55, Issue 3, 9 November 2020, Pages 298-313.e10.
\href{https://www.cell.com/developmental-cell/pdfExtended/S1534-5807(20)30796-6}{doi.org/10.1016/j.devcel.2020.10.006}.
}
\usage{
peptides.raw
}
\value{
A data.frame.
}
\description{
The dataset contains the peptides.txt  output file from
\href{http://www.maxquant.org}{MaxQuant}.
}
\keyword{datasets}
