% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_labelingefficiency.R
\name{plot_labelingefficiency}
\alias{plot_labelingefficiency}
\title{Plot TMT labeling efficiency}
\usage{
plot_labelingefficiency(evidence)
}
\arguments{
\item{evidence}{Data.frame,
evidence table originating from MaxQuant.}
}
\value{
A scatter plot (generated by \code{\link[ggplot2]{ggplot}})
}
\description{
\code{plot_labelingefficiency} generates a stacked barplot for peptides
modified by TMT at the N-termini and Lys residues.
}
\examples{
if(interactive()){
data("evidence.raw")
evidencedata <- evidence.raw
plot_labelingefficiency(evidencedata)
}
}
