% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_miscleavagerate.R
\name{plot_miscleavagerate}
\alias{plot_miscleavagerate}
\title{plot_miscleavagerate}
\usage{
plot_miscleavagerate(peptides)
}
\arguments{
\item{peptides}{Data.frame,
Peptide table originating from MaxQuant.}
}
\value{
A scatter plot (generated by \code{\link[ggplot2]{ggplot}})
}
\description{
plot_miscleavagerate
}
\examples{
#' if(interactive()){
data("peptides.raw")
pepdata <- peptides.raw
plot_miscleavagerate(pepdata)
}
}
